+++ event-sources/event-source-kafka/src/main/java/io/telicent/smart/cache/sources/kafka/KafkaDatasetGraphSource.java
@@ -67,10 +67,11 @@ public class KafkaDatasetGraphSource<TKey> extends KafkaEventSource<TKey, Datase
      */
     KafkaDatasetGraphSource(String bootstrapServers, Set<String> topics, String groupId, String keyDeserializerClass,
                             int maxPollRecords, KafkaReadPolicy<TKey, DatasetGraph> policy, boolean autoCommit,
-                            OffsetStore offsetStore, Duration lagReportInterval, Properties properties) {
+                            boolean ignoreTombstones, OffsetStore offsetStore, Duration lagReportInterval,
+                            Properties properties) {
         super(bootstrapServers, topics, groupId, keyDeserializerClass,
-              DatasetGraphDeserializer.class.getCanonicalName(), maxPollRecords, policy, autoCommit, offsetStore,
-              lagReportInterval, properties);
+              DatasetGraphDeserializer.class.getCanonicalName(), maxPollRecords, policy, autoCommit, ignoreTombstones,
+              offsetStore, lagReportInterval, properties);
     }
 
     /**
@@ -92,8 +93,8 @@ public class KafkaDatasetGraphSource<TKey> extends KafkaEventSource<TKey, Datase
         public KafkaDatasetGraphSource<TKey> build() {
             return new KafkaDatasetGraphSource<>(this.bootstrapServers, this.topics, this.groupId,
                                                  this.keyDeserializerClass, this.maxPollRecords, this.readPolicy,
-                                                 this.autoCommit, this.externalOffsetStore, this.lagReportInterval,
-                                                 this.properties);
+                                                 this.autoCommit, this.ignoreTombstones, this.externalOffsetStore,
+                                                 this.lagReportInterval, this.properties);
         }
     }
 }