+++ event-sources/event-source-kafka/src/main/java/io/telicent/smart/cache/sources/kafka/KafkaEventSource.java
@@ -25,6 +25,7 @@ import io.telicent.smart.cache.observability.TelicentMetrics;
 import io.telicent.smart.cache.projectors.utils.PeriodicAction;
 import io.telicent.smart.cache.sources.Event;
 import io.telicent.smart.cache.sources.EventSourceException;
+import io.telicent.smart.cache.sources.buffered.AbstractBufferedEventSource;
 import io.telicent.smart.cache.sources.kafka.policies.KafkaReadPolicy;
 import io.telicent.smart.cache.sources.offsets.OffsetStore;
 import org.apache.commons.collections4.CollectionUtils;
@@ -86,7 +87,7 @@ public class KafkaEventSource<TKey, TValue>
     private final Set<String> topics;
     private boolean firstRun = true;
     private final TopicExistenceChecker topicExistenceChecker;
-    private final boolean autoCommit;
+    private final boolean autoCommit, ignoreTombstones;
     private final Map<TopicPartition, OffsetAndMetadata> autoCommitOffsets = new HashMap<>();
     private final Queue<Map<TopicPartition, OffsetAndMetadata>> delayedOffsetCommits = new ConcurrentLinkedDeque<>();
     private final Map<TopicPartition, Long> delayedOffsetResets = new ConcurrentHashMap<>();
@@ -114,6 +115,7 @@ public class KafkaEventSource<TKey, TValue>
      *                               {@link KafkaConsumer#poll(Duration)} request.
      * @param policy                 Kafka Read Policy to control what events to read from the configured topic
      * @param autoCommit             Whether the event source will automatically commit Kafka positions
+     * @param ignoreTombstones       Whether to ignore tombstone events and not return them from poll() calls
      * @param offsetStore            An external offset store to commit offsets to in addition to committing them to
      *                               Kafka
      * @param lagReportInterval      Lag reporting interval
@@ -123,7 +125,8 @@ public class KafkaEventSource<TKey, TValue>
     KafkaEventSource(final String bootstrapServers, final Set<String> topics, final String groupId,
                      final String keyDeserializerClass, final String valueDeserializerClass, final int maxPollRecords,
                      final KafkaReadPolicy<TKey, TValue> policy, final boolean autoCommit,
-                     final OffsetStore offsetStore, final Duration lagReportInterval, final Properties properties) {
+                     final boolean ignoreTombstones, final OffsetStore offsetStore, final Duration lagReportInterval,
+                     final Properties properties) {
         if (StringUtils.isBlank(bootstrapServers)) {
             throw new IllegalArgumentException("Kafka bootstrapServers cannot be null");
         }
@@ -170,6 +173,7 @@ public class KafkaEventSource<TKey, TValue>
         this.readPolicy = policy;
         this.readPolicy.setConsumer(this.consumer);
         this.autoCommit = autoCommit;
+        this.ignoreTombstones = ignoreTombstones;
         this.externalOffsetStore = offsetStore;
         this.topicExistenceChecker =
                 new TopicExistenceChecker(createAdminClient(props), this.server, this.topics, LOGGER);
@@ -392,8 +396,7 @@ public class KafkaEventSource<TKey, TValue>
      * @see io.telicent.smart.cache.sources.EventSource#processed(Collection)
      */
     @Override
-    @SuppressWarnings("rawtypes")
-    public void processed(Collection<Event> processedEvents) {
+    public void processed(Collection<Event<?,?>> processedEvents) {
         // Compute the maximum processed offset for each topic partitions
         Map<TopicPartition, OffsetAndMetadata> commitOffsets = determineCommitOffsetsFromEvents(processedEvents);
 
@@ -467,7 +470,7 @@ public class KafkaEventSource<TKey, TValue>
      * @return Offsets to commit, may be empty if no Kafka events provided
      */
     @SuppressWarnings("rawtypes")
-    public static Map<TopicPartition, OffsetAndMetadata> determineCommitOffsetsFromEvents(Collection<Event> events) {
+    public static Map<TopicPartition, OffsetAndMetadata> determineCommitOffsetsFromEvents(Collection<Event<?, ?>> events) {
         if (CollectionUtils.isEmpty(events)) {
             return Collections.emptyMap();
         }
@@ -591,30 +594,20 @@ public class KafkaEventSource<TKey, TValue>
         return false;
     }
 
-    private static Duration updateTimeout(long start, Duration timeout) {
-        long elapsed = System.currentTimeMillis() - start;
-        long remainingTime = timeout.toMillis() - elapsed;
-        if (remainingTime <= 0) {
-            return null;
-        } else {
-            return Duration.ofMillis(remainingTime);
-        }
-    }
-
     @Override
-    protected void tryFillBuffer(Duration timeout) {
+    protected boolean tryFillBuffer(Duration timeout) {
         // Buffer up some more events
         ConsumerRecords<TKey, TValue> records;
         try {
             // Don't do any work if none of the topic(s) exist on the Kafka cluster
             long start = System.currentTimeMillis();
-            if (!this.topicExistenceChecker.anyTopicExists(timeout)) return;
+            if (!this.topicExistenceChecker.anyTopicExists(timeout)) return true;
 
             // Reduce the timeout by the amount of time we spent waiting for the topic to exist as otherwise we
             // could wait twice our timeout and violate our API contract
             timeout = updateTimeout(start, timeout);
             if (timeout == null) {
-                return;
+                return false;
             }
 
             // If an offset reset is in progress this means we have delayed offset resets submitted by a thread other
@@ -636,7 +629,7 @@ public class KafkaEventSource<TKey, TValue>
                 // In this scenario the events we just returned are likely not for the correct offsets so don't
                 // add these to the buffer, this forces the caller to call EventSource.poll() again at which point we
                 // should resolve the offset reset on our next poll() call and return the expected events.
-                return;
+                return false;
             }
             for (ConsumerRecord<TKey, TValue> record : records) {
                 events.add(record);
@@ -650,14 +643,38 @@ public class KafkaEventSource<TKey, TValue>
                              topicNames);
             }
 
+            // Proactively prune tombstones
+            // We do this here rather than in decodeEvent() as this allows us to potentially discard the entire buffer
+            // if we've hit a segment of the topic with a lot of tombstones, this causes tryFillBuffer() to get called
+            // again and retrieve the next batch of events sooner
+            // Note whether we retrieved anything or not prior to this check as if we eliminate all events due to them
+            // all being tombstones we want poll() to call us again promptly
+            boolean genuinelyEmpty = events.isEmpty();
+            if (this.ignoreTombstones) {
+                int size = events.size();
+                events.removeIf(r -> r.value() == null);
+                if (events.size() < size) {
+                    LOGGER.debug("[{}] Pruned {} tombstone events retrieved from Kafka", topicNames,
+                                 size - events.size());
+                }
+            }
+
             this.positionLogger.run();
 
+            // Return true if, and only if, we retrieved zero new events from Kafka indicating nothing was currently
+            // available
+            return genuinelyEmpty;
         /*
         These errors are considered recoverable i.e. a subsequent call to this function could successfully fill the
         buffer
         */
         } catch (WakeupException | InterruptException e) {
             LOGGER.debug("[{}] Interrupted/woken while polling Kafka for events", this.topicNames);
+
+            // Returning true to indicate we were interrupted, this prevents poll() from immediately calling us again
+            // and ensures that interrupts propagate upwards promptly as it's likely higher level code will also detect
+            // the interrupt and stop poll()'ing
+            return true;
         /*
         The following errors are considered unrecoverable and result in an EventSourceException being thrown
 
@@ -714,7 +731,8 @@ public class KafkaEventSource<TKey, TValue>
         boolean anyTopicsMatch = offsets.keySet().stream().anyMatch(t -> this.topics.contains(t.topic()));
         if (!anyTopicsMatch) {
             LOGGER.warn(
-                    "[{}] Reset offsets called without any topics that match this sources configured topics, nothing was reset!", this.topicNames);
+                    "[{}] Reset offsets called without any topics that match this sources configured topics, nothing was reset!",
+                    this.topicNames);
             return;
         }
 
@@ -729,7 +747,8 @@ public class KafkaEventSource<TKey, TValue>
         if (this.pollThread != Thread.currentThread()) {
             // Delay resets to later
             this.delayedOffsetResets.putAll(offsets);
-            LOGGER.info("[{}] Only the polling thread may reset offsets, delaying resets until next poll call", this.topicNames);
+            LOGGER.info("[{}] Only the polling thread may reset offsets, delaying resets until next poll call",
+                        this.topicNames);
 
             // Clear out internally buffered events, this forces the next poll() call to call tryFillBuffer() which is
             // where we process our delayed offset resets
@@ -805,8 +824,8 @@ public class KafkaEventSource<TKey, TValue>
         public KafkaEventSource<TKey, TValue> build() {
             return new KafkaEventSource<>(this.bootstrapServers, this.topics, this.groupId, this.keyDeserializerClass,
                                           this.valueDeserializerClass, this.maxPollRecords, this.readPolicy,
-                                          this.autoCommit, this.externalOffsetStore, this.lagReportInterval,
-                                          this.properties);
+                                          this.autoCommit, this.ignoreTombstones, this.externalOffsetStore,
+                                          this.lagReportInterval, this.properties);
         }
     }