+++ event-sources/event-source-kafka/src/test/java/io/telicent/smart/cache/sources/kafka/SecureKafkaTestCluster.java
@@ -170,6 +170,6 @@ public class SecureKafkaTestCluster extends BasicKafkaTestCluster {
     @Override
     protected int getDefaultTimeout() {
         // NB - Takes slightly longer to connect to secure Kafka clusters
-        return 5;
+        return super.getDefaultTimeout() + 5;
     }
 }