+++ event-sources/event-source-kafka/src/test/java/io/telicent/smart/cache/sources/kafka/tombstones/DataDeserializer.java
@@ -13,19 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.telicent.smart.cache.sources.kafka.serializers;
+package io.telicent.smart.cache.sources.kafka.tombstones;
 
-import java.util.Map;
+import io.telicent.smart.cache.sources.kafka.serializers.AbstractJacksonDeserializer;
 
-public class TestPayloadDeserializerBadConfig extends TestPayloadDeserializer {
+public class DataDeserializer extends AbstractJacksonDeserializer<Data> {
 
-    private final Map<String, ?> BAD_PARSING_CONFIGURATION =
-            Map.of(RdfPayloadDeserializer.EAGER_PARSING_CONFIG_KEY, 1234);
-
-    @Override
-    protected RdfPayloadDeserializer createPayloadDeserializer() {
-        RdfPayloadDeserializer deserializer = super.createPayloadDeserializer();
-        deserializer.configure(BAD_PARSING_CONFIGURATION, false);
-        return deserializer;
+    public DataDeserializer() {
+        super(Data.class);
     }
 }